Initial Commit
This initial commit will include a functioning Object Oriented Login System designed and built in PHP.

Various changes will have to be implemented before V1.0 releases.

This release is still in Beta testing, and some functionality has not been included such as User Database Management, Design Changes, and more.

If there is an issue with any of the code, please submit an Issue, and give as much information as you can.