Fixed Bug
When creating the Session::flash() Update, the code was broken because most of the Session::flash() call's were only defining 2 arguments, and the Session token was using session::put() to which both flash() and put() had to define type as an empty string to fix. Also through in an If Else statement.