Database Management Initial Commit + Delete Table Functionality
Added the Database Management portion of the Admin section of the website.

Listed the tables on the Database Management "home page". Made sure that the tables "users", "user_session", and "groups" was not able to be edited, or deleted.

Cross checked these tables in the confirmation page and the deletion page.

Also added the functionality to delete a table from the database if it exists.