Added Google+ Login API
Added the Google+ Login API to the website. This is version 1 of this branch. The updates will include changing the "login with google" to a "link your google account" page. That way we can grab up to date information if a User has a google account. Their account e-mail must match the google account's email address. So we have to add a "Change Email" section of the settings page to make sure that the user has the Ability to change this information if needed.

Otherwise the functionality of this update includes:

Logging in with google
Logging out of google account
When logging in, checking to see if the user has linked their google account with our database before, if not, update the database.
If the user isn't in our system at all, force them to register first with out system

Enjoy the update! ;)