Breadcrumbs And More...
Added Breadcrumbs into the mix and out came a wonderful piece of meatloaf. :P Well, maybe not meatloaf but it certainly made things easier for navigation!

Also changed the folder structure of the admin backpage. Having the user not be able to type /admin/usermanage into the url was not ideal. Updated to compensate for not doing it this way initially