Bug Fix: Settings Page
Fixed a bug on the settings page where the z-index of the nav bar was not set properly. This was causing a bug where the user would sometimes see the navbar dropdown menu behind an object within the body... i.e. the iframe. Set the z-index of the nav bar at 100, so that any page will always have the nav bar on the top.