Created The Team Page!
With this commit we are now including "The Team" page. I decided to change this from "Members" as I honestly don't think it was be smart to list all the users out for various reasons. Instead to solve this influx of functionality we will be adding a search bar (In the navbar) powered by jquery to search for any user on command.

Welcome to The Team Page!