Initial Commit to Branch Search-Functionality
In this commit, the search input field is placed in the NavBar with functional javascript input validation in the navbar.js file.

Also added a dropdown where the user can change where he or she would like to search in the website.  This is essentially whish "table" the user wants to search. The User's Link is going to be the default choice. The type is passed through a hidden input with the value being handled by javascript on click functions.

Also changed the way animate.css was loading by adding it to the composer file. This is going to be an ongoing effort to load all the dependancies into the composer file if possible.

Also started the Search Class file getting ready for the upcoming events in this branch.